Dramatic refactoring of VSC85xx driver
