+++ drv/vsc85xx/src/tesla.rs
@@ -1,0 +1,442 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, self. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+use core::ops::{Deref, DerefMut};
+
+use crate::Trace;
+use crate::{Phy, PhyRw};
+use ringbuf::ringbuf_entry_root as ringbuf_entry;
+use vsc7448_pac::phy;
+use vsc_err::VscError;
+
+pub struct TeslaPhy<'a, 'b, P>(pub &'b mut Phy<'a, P>);
+impl<'a, 'b, P> Deref for TeslaPhy<'a, 'b, P> {
+    type Target = Phy<'a, P>;
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl<'a, 'b, P> DerefMut for TeslaPhy<'a, 'b, P> {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.0
+    }
+}
+
+impl<'a, 'b, P: PhyRw> TeslaPhy<'a, 'b, P> {
+    /// Applies a patch to the 8051 microcode inside the PHY, based on
+    /// `vtss_phy_pre_init_seq_tesla_rev_e` in the SDK
+    pub(crate) fn patch(&mut self) -> Result<(), VscError> {
+        ringbuf_entry!(Trace::TeslaPatch(self.port));
+
+        // Enable broadcast flag to configure all ports simultaneously
+        self.modify(phy::STANDARD::EXTENDED_CONTROL_AND_STATUS(), |r| {
+            *r = (u16::from(*r) | 1).into();
+        })?;
+
+        self.write(phy::STANDARD::EXTENDED_PHY_CONTROL_2(), 0x0040.into())?;
+        self.write(phy::EXTENDED_2::CU_PMD_TX_CTRL(), 0x02be.into())?;
+        self.write(phy::TEST::TEST_PAGE_20(), 0x4320.into())?;
+        self.write(phy::TEST::TEST_PAGE_24(), 0x0c00.into())?;
+        self.write(phy::TEST::TEST_PAGE_9(), 0x18ca.into())?;
+        self.write(phy::TEST::TEST_PAGE_5(), 0x1b20.into())?;
+
+        // "Enable token-ring during coma-mode"
+        self.modify(phy::TEST::TEST_PAGE_8(), |r| {
+            r.0 |= 0x8000;
+        })?;
+
+        self.write(phy::TR::TR_18(), 0x0004.into())?;
+        self.write(phy::TR::TR_17(), 0x01bd.into())?;
+        self.write(phy::TR::TR_16(), 0x8fae.into())?;
+        self.write(phy::TR::TR_18(), 0x000f.into())?;
+        self.write(phy::TR::TR_17(), 0x000f.into())?;
+        self.write(phy::TR::TR_16(), 0x8fac.into())?;
+        self.write(phy::TR::TR_18(), 0x00a0.into())?;
+        self.write(phy::TR::TR_17(), 0xf147.into())?;
+        self.write(phy::TR::TR_16(), 0x97a0.into())?;
+        self.write(phy::TR::TR_18(), 0x0005.into())?;
+        self.write(phy::TR::TR_17(), 0x2f54.into())?;
+        self.write(phy::TR::TR_16(), 0x8fe4.into())?;
+        self.write(phy::TR::TR_18(), 0x0027.into())?;
+        self.write(phy::TR::TR_17(), 0x303d.into())?;
+        self.write(phy::TR::TR_16(), 0x9792.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x0704.into())?;
+        self.write(phy::TR::TR_16(), 0x87fe.into())?;
+        self.write(phy::TR::TR_18(), 0x0006.into())?;
+        self.write(phy::TR::TR_17(), 0x0150.into())?;
+        self.write(phy::TR::TR_16(), 0x8fe0.into())?;
+        self.write(phy::TR::TR_18(), 0x0012.into())?;
+        self.write(phy::TR::TR_17(), 0xb00a.into())?;
+        self.write(phy::TR::TR_16(), 0x8f82.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x0d74.into())?;
+        self.write(phy::TR::TR_16(), 0x8f80.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x0012.into())?;
+        self.write(phy::TR::TR_16(), 0x82e0.into())?;
+        self.write(phy::TR::TR_18(), 0x0005.into())?;
+        self.write(phy::TR::TR_17(), 0x0208.into())?;
+        self.write(phy::TR::TR_16(), 0x83a2.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x9186.into())?;
+        self.write(phy::TR::TR_16(), 0x83b2.into())?;
+        self.write(phy::TR::TR_18(), 0x000e.into())?;
+        self.write(phy::TR::TR_17(), 0x3700.into())?;
+        self.write(phy::TR::TR_16(), 0x8fb0.into())?;
+        self.write(phy::TR::TR_18(), 0x0004.into())?;
+        self.write(phy::TR::TR_17(), 0x9f81.into())?;
+        self.write(phy::TR::TR_16(), 0x9688.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0xffff.into())?;
+        self.write(phy::TR::TR_16(), 0x8fd2.into())?;
+        self.write(phy::TR::TR_18(), 0x0003.into())?;
+        self.write(phy::TR::TR_17(), 0x9fa2.into())?;
+        self.write(phy::TR::TR_16(), 0x968a.into())?;
+        self.write(phy::TR::TR_18(), 0x0020.into())?;
+        self.write(phy::TR::TR_17(), 0x640b.into())?;
+        self.write(phy::TR::TR_16(), 0x9690.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x2220.into())?;
+        self.write(phy::TR::TR_16(), 0x8258.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x2a20.into())?;
+        self.write(phy::TR::TR_16(), 0x825a.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x3060.into())?;
+        self.write(phy::TR::TR_16(), 0x825c.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x3fa0.into())?;
+        self.write(phy::TR::TR_16(), 0x825e.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0xe0f0.into())?;
+        self.write(phy::TR::TR_16(), 0x83a6.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x1489.into())?;
+        self.write(phy::TR::TR_16(), 0x8f92.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x7000.into())?;
+        self.write(phy::TR::TR_16(), 0x96a2.into())?;
+        self.write(phy::TR::TR_18(), 0x0007.into())?;
+        self.write(phy::TR::TR_17(), 0x1448.into())?;
+        self.write(phy::TR::TR_16(), 0x96a6.into())?;
+        self.write(phy::TR::TR_18(), 0x00ee.into())?;
+        self.write(phy::TR::TR_17(), 0xffdd.into())?;
+        self.write(phy::TR::TR_16(), 0x96a0.into())?;
+        self.write(phy::TR::TR_18(), 0x0091.into())?;
+        self.write(phy::TR::TR_17(), 0xb06c.into())?;
+        self.write(phy::TR::TR_16(), 0x8fe8.into())?;
+        self.write(phy::TR::TR_18(), 0x0004.into())?;
+        self.write(phy::TR::TR_17(), 0x1600.into())?;
+        self.write(phy::TR::TR_16(), 0x8fea.into())?;
+        self.write(phy::TR::TR_18(), 0x00ee.into())?;
+        self.write(phy::TR::TR_17(), 0xff00.into())?;
+        self.write(phy::TR::TR_16(), 0x96b0.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x7000.into())?;
+        self.write(phy::TR::TR_16(), 0x96b2.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x0814.into())?;
+        self.write(phy::TR::TR_16(), 0x96b4.into())?;
+        self.write(phy::TR::TR_18(), 0x0068.into())?;
+        self.write(phy::TR::TR_17(), 0x8980.into())?;
+        self.write(phy::TR::TR_16(), 0x8f90.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0xd8f0.into())?;
+        self.write(phy::TR::TR_16(), 0x83a4.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x0400.into())?;
+        self.write(phy::TR::TR_16(), 0x8fc0.into())?;
+        self.write(phy::TR::TR_18(), 0x0050.into())?;
+        self.write(phy::TR::TR_17(), 0x100f.into())?;
+        self.write(phy::TR::TR_16(), 0x87fa.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x0003.into())?;
+        self.write(phy::TR::TR_16(), 0x8796.into())?;
+        self.write(phy::TR::TR_18(), 0x00c3.into())?;
+        self.write(phy::TR::TR_17(), 0xff98.into())?;
+        self.write(phy::TR::TR_16(), 0x87f8.into())?;
+        self.write(phy::TR::TR_18(), 0x0018.into())?;
+        self.write(phy::TR::TR_17(), 0x292a.into())?;
+        self.write(phy::TR::TR_16(), 0x8fa4.into())?;
+        self.write(phy::TR::TR_18(), 0x00d2.into())?;
+        self.write(phy::TR::TR_17(), 0xc46f.into())?;
+        self.write(phy::TR::TR_16(), 0x968c.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x0620.into())?;
+        self.write(phy::TR::TR_16(), 0x97a2.into())?;
+        self.write(phy::TR::TR_18(), 0x0013.into())?;
+        self.write(phy::TR::TR_17(), 0x132f.into())?;
+        self.write(phy::TR::TR_16(), 0x96a4.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x0000.into())?;
+        self.write(phy::TR::TR_16(), 0x96a8.into())?;
+        self.write(phy::TR::TR_18(), 0x00c0.into())?;
+        self.write(phy::TR::TR_17(), 0xa028.into())?;
+        self.write(phy::TR::TR_16(), 0x8ffc.into())?;
+        self.write(phy::TR::TR_18(), 0x0090.into())?;
+        self.write(phy::TR::TR_17(), 0x1c09.into())?;
+        self.write(phy::TR::TR_16(), 0x8fec.into())?;
+        self.write(phy::TR::TR_18(), 0x0004.into())?;
+        self.write(phy::TR::TR_17(), 0xa6a1.into())?;
+        self.write(phy::TR::TR_16(), 0x8fee.into())?;
+        self.write(phy::TR::TR_18(), 0x00b0.into())?;
+        self.write(phy::TR::TR_17(), 0x1807.into())?;
+        self.write(phy::TR::TR_16(), 0x8ffe.into())?;
+
+        // We're not using 10BASE-TE, so this is the correct config block
+        self.write(phy::TR::TR_16(), 0x028e.into())?;
+        self.write(phy::TR::TR_18(), 0x0008.into())?;
+        self.write(phy::TR::TR_17(), 0xa518.into())?;
+        self.write(phy::TR::TR_16(), 0x8486.into())?;
+        self.write(phy::TR::TR_18(), 0x006d.into())?;
+        self.write(phy::TR::TR_17(), 0xc696.into())?;
+        self.write(phy::TR::TR_16(), 0x8488.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x0912.into())?;
+        self.write(phy::TR::TR_16(), 0x848a.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x0db6.into())?;
+        self.write(phy::TR::TR_16(), 0x848e.into())?;
+        self.write(phy::TR::TR_18(), 0x0059.into())?;
+        self.write(phy::TR::TR_17(), 0x6596.into())?;
+        self.write(phy::TR::TR_16(), 0x849c.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x0514.into())?;
+        self.write(phy::TR::TR_16(), 0x849e.into())?;
+        self.write(phy::TR::TR_18(), 0x0041.into())?;
+        self.write(phy::TR::TR_17(), 0x0280.into())?;
+        self.write(phy::TR::TR_16(), 0x84a2.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x0000.into())?;
+        self.write(phy::TR::TR_16(), 0x84a4.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x0000.into())?;
+        self.write(phy::TR::TR_16(), 0x84a6.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x0000.into())?;
+        self.write(phy::TR::TR_16(), 0x84a8.into())?;
+        self.write(phy::TR::TR_18(), 0x0000.into())?;
+        self.write(phy::TR::TR_17(), 0x0000.into())?;
+        self.write(phy::TR::TR_16(), 0x84aa.into())?;
+        self.write(phy::TR::TR_18(), 0x007d.into())?;
+        self.write(phy::TR::TR_17(), 0xf7dd.into())?;
+        self.write(phy::TR::TR_16(), 0x84ae.into())?;
+        self.write(phy::TR::TR_18(), 0x006d.into())?;
+        self.write(phy::TR::TR_17(), 0x95d4.into())?;
+        self.write(phy::TR::TR_16(), 0x84b0.into())?;
+        self.write(phy::TR::TR_18(), 0x0049.into())?;
+        self.write(phy::TR::TR_17(), 0x2410.into())?;
+        self.write(phy::TR::TR_16(), 0x84b2.into())?;
+
+        self.modify(phy::TEST::TEST_PAGE_8(), |r| {
+            r.0 &= !0x8000; // Disable token-ring mode
+        })?;
+
+        self.modify(phy::STANDARD::EXTENDED_CONTROL_AND_STATUS(), |r| {
+            *r = (u16::from(*r) & !1).into();
+        })?;
+
+        //////////////////////////////////////////////////////////////////////////
+        // Now we're going deep into the weeds.  This section is based on
+        // `tesla_revB_8051_patch` in the SDK, which (as the name suggests), patches
+        // the 8051 in the PHY.
+        const FIRMWARE_START_ADDR: u16 = 0x4000;
+        const PATCH_CRC_LEN: u16 = (TESLA_PATCH.len() + 1) as u16;
+        const EXPECTED_CRC: u16 = 0x29E8;
+
+        // This patch can only be applied to Port 0 of the PHY, so we'll check
+        // the address here.
+        let phy_port =
+            self.read(phy::EXTENDED::EXTENDED_PHY_CONTROL_4())?.0 >> 11;
+        if phy_port != 0 {
+            return Err(VscError::BadPhyPatchPort(phy_port));
+        }
+        let crc = self.read_8051_crc(FIRMWARE_START_ADDR, PATCH_CRC_LEN)?;
+        let skip_download = crc == EXPECTED_CRC;
+        let patch_ok = skip_download
+            && self.read(phy::GPIO::GPIO_3())?.0 == 0x3eb7
+            && self.read(phy::GPIO::GPIO_4())?.0 == 0x4012
+            && self.read(phy::GPIO::GPIO_12())?.0 == 0x0100
+            && self.read(phy::GPIO::GPIO_0())?.0 == 0xc018;
+
+        ringbuf_entry!(Trace::PatchState {
+            patch_ok,
+            skip_download
+        });
+
+        if !skip_download || !patch_ok {
+            self.micro_assert_reset()?;
+        }
+        if !skip_download {
+            self.download_patch(&TESLA_PATCH)?;
+        }
+        if !patch_ok {
+            // Various CPU commands to enable the patch
+            self.write(phy::GPIO::GPIO_3(), 0x3eb7.into())?;
+            self.write(phy::GPIO::GPIO_4(), 0x4012.into())?;
+            self.write(phy::GPIO::GPIO_12(), 0x0100.into())?;
+            self.write(phy::GPIO::GPIO_0(), 0xc018.into())?;
+        }
+
+        if !skip_download {
+            let crc = self.read_8051_crc(FIRMWARE_START_ADDR, PATCH_CRC_LEN)?;
+            if crc != EXPECTED_CRC {
+                return Err(VscError::PhyPatchFailedCrc);
+            }
+        }
+
+        //////////////////////////////////////////////////////////////////////////
+        // `vtss_phy_pre_init_tesla_revB_1588`
+        //
+        // "Pass the cmd to Micro to initialize all 1588 analyzer registers to
+        //  default"
+        self.cmd(0x801A)?;
+
+        Ok(())
+    }
+}
+
+/// Raw patch for 8051 microcode, from `tesla_revB_8051_patch` in the SDK
+const TESLA_PATCH: [u8; 1655] = [
+    0x46, 0x4a, 0x02, 0x43, 0x37, 0x02, 0x46, 0x26, 0x02, 0x46, 0x77, 0x02,
+    0x45, 0x60, 0x02, 0x45, 0xaf, 0xed, 0xff, 0xe5, 0xfc, 0x54, 0x38, 0x64,
+    0x20, 0x70, 0x08, 0x65, 0xff, 0x70, 0x04, 0xed, 0x44, 0x80, 0xff, 0x22,
+    0x8f, 0x19, 0x7b, 0xbb, 0x7d, 0x0e, 0x7f, 0x04, 0x12, 0x3d, 0xd7, 0xef,
+    0x4e, 0x60, 0x03, 0x02, 0x41, 0xf9, 0xe4, 0xf5, 0x1a, 0x74, 0x01, 0x7e,
+    0x00, 0xa8, 0x1a, 0x08, 0x80, 0x05, 0xc3, 0x33, 0xce, 0x33, 0xce, 0xd8,
+    0xf9, 0xff, 0xef, 0x55, 0x19, 0x70, 0x03, 0x02, 0x41, 0xed, 0x85, 0x1a,
+    0xfb, 0x7b, 0xbb, 0xe4, 0xfd, 0xff, 0x12, 0x3d, 0xd7, 0xef, 0x4e, 0x60,
+    0x03, 0x02, 0x41, 0xed, 0xe5, 0x1a, 0x54, 0x02, 0x75, 0x1d, 0x00, 0x25,
+    0xe0, 0x25, 0xe0, 0xf5, 0x1c, 0xe4, 0x78, 0xc5, 0xf6, 0xd2, 0x0a, 0x12,
+    0x41, 0xfa, 0x7b, 0xff, 0x7d, 0x12, 0x7f, 0x07, 0x12, 0x3d, 0xd7, 0xef,
+    0x4e, 0x60, 0x03, 0x02, 0x41, 0xe7, 0xc2, 0x0a, 0x74, 0xc7, 0x25, 0x1a,
+    0xf9, 0x74, 0xe7, 0x25, 0x1a, 0xf8, 0xe6, 0x27, 0xf5, 0x1b, 0xe5, 0x1d,
+    0x24, 0x5b, 0x12, 0x45, 0xea, 0x12, 0x3e, 0xda, 0x7b, 0xfc, 0x7d, 0x11,
+    0x7f, 0x07, 0x12, 0x3d, 0xd7, 0x78, 0xcc, 0xef, 0xf6, 0x78, 0xc1, 0xe6,
+    0xfe, 0xef, 0xd3, 0x9e, 0x40, 0x06, 0x78, 0xcc, 0xe6, 0x78, 0xc1, 0xf6,
+    0x12, 0x41, 0xfa, 0x7b, 0xec, 0x7d, 0x12, 0x7f, 0x07, 0x12, 0x3d, 0xd7,
+    0x78, 0xcb, 0xef, 0xf6, 0xbf, 0x07, 0x06, 0x78, 0xc3, 0x76, 0x1a, 0x80,
+    0x1f, 0x78, 0xc5, 0xe6, 0xff, 0x60, 0x0f, 0xc3, 0xe5, 0x1b, 0x9f, 0xff,
+    0x78, 0xcb, 0xe6, 0x85, 0x1b, 0xf0, 0xa4, 0x2f, 0x80, 0x07, 0x78, 0xcb,
+    0xe6, 0x85, 0x1b, 0xf0, 0xa4, 0x78, 0xc3, 0xf6, 0xe4, 0x78, 0xc2, 0xf6,
+    0x78, 0xc2, 0xe6, 0xff, 0xc3, 0x08, 0x96, 0x40, 0x03, 0x02, 0x41, 0xd1,
+    0xef, 0x54, 0x03, 0x60, 0x33, 0x14, 0x60, 0x46, 0x24, 0xfe, 0x60, 0x42,
+    0x04, 0x70, 0x4b, 0xef, 0x24, 0x02, 0xff, 0xe4, 0x33, 0xfe, 0xef, 0x78,
+    0x02, 0xce, 0xa2, 0xe7, 0x13, 0xce, 0x13, 0xd8, 0xf8, 0xff, 0xe5, 0x1d,
+    0x24, 0x5c, 0xcd, 0xe5, 0x1c, 0x34, 0xf0, 0xcd, 0x2f, 0xff, 0xed, 0x3e,
+    0xfe, 0x12, 0x46, 0x0d, 0x7d, 0x11, 0x80, 0x0b, 0x78, 0xc2, 0xe6, 0x70,
+    0x04, 0x7d, 0x11, 0x80, 0x02, 0x7d, 0x12, 0x7f, 0x07, 0x12, 0x3e, 0x9a,
+    0x8e, 0x1e, 0x8f, 0x1f, 0x80, 0x03, 0xe5, 0x1e, 0xff, 0x78, 0xc5, 0xe6,
+    0x06, 0x24, 0xcd, 0xf8, 0xa6, 0x07, 0x78, 0xc2, 0x06, 0xe6, 0xb4, 0x1a,
+    0x0a, 0xe5, 0x1d, 0x24, 0x5c, 0x12, 0x45, 0xea, 0x12, 0x3e, 0xda, 0x78,
+    0xc5, 0xe6, 0x65, 0x1b, 0x70, 0x82, 0x75, 0xdb, 0x20, 0x75, 0xdb, 0x28,
+    0x12, 0x46, 0x02, 0x12, 0x46, 0x02, 0xe5, 0x1a, 0x12, 0x45, 0xf5, 0xe5,
+    0x1a, 0xc3, 0x13, 0x12, 0x45, 0xf5, 0x78, 0xc5, 0x16, 0xe6, 0x24, 0xcd,
+    0xf8, 0xe6, 0xff, 0x7e, 0x08, 0x1e, 0xef, 0xa8, 0x06, 0x08, 0x80, 0x02,
+    0xc3, 0x13, 0xd8, 0xfc, 0xfd, 0xc4, 0x33, 0x54, 0xe0, 0xf5, 0xdb, 0xef,
+    0xa8, 0x06, 0x08, 0x80, 0x02, 0xc3, 0x13, 0xd8, 0xfc, 0xfd, 0xc4, 0x33,
+    0x54, 0xe0, 0x44, 0x08, 0xf5, 0xdb, 0xee, 0x70, 0xd8, 0x78, 0xc5, 0xe6,
+    0x70, 0xc8, 0x75, 0xdb, 0x10, 0x02, 0x40, 0xfd, 0x78, 0xc2, 0xe6, 0xc3,
+    0x94, 0x17, 0x50, 0x0e, 0xe5, 0x1d, 0x24, 0x62, 0x12, 0x42, 0x08, 0xe5,
+    0x1d, 0x24, 0x5c, 0x12, 0x42, 0x08, 0x20, 0x0a, 0x03, 0x02, 0x40, 0x76,
+    0x05, 0x1a, 0xe5, 0x1a, 0xc3, 0x94, 0x04, 0x50, 0x03, 0x02, 0x40, 0x3a,
+    0x22, 0xe5, 0x1d, 0x24, 0x5c, 0xff, 0xe5, 0x1c, 0x34, 0xf0, 0xfe, 0x12,
+    0x46, 0x0d, 0x22, 0xff, 0xe5, 0x1c, 0x34, 0xf0, 0xfe, 0x12, 0x46, 0x0d,
+    0x22, 0xe4, 0xf5, 0x19, 0x12, 0x46, 0x43, 0x20, 0xe7, 0x1e, 0x7b, 0xfe,
+    0x12, 0x42, 0xf9, 0xef, 0xc4, 0x33, 0x33, 0x54, 0xc0, 0xff, 0xc0, 0x07,
+    0x7b, 0x54, 0x12, 0x42, 0xf9, 0xd0, 0xe0, 0x4f, 0xff, 0x74, 0x2a, 0x25,
+    0x19, 0xf8, 0xa6, 0x07, 0x12, 0x46, 0x43, 0x20, 0xe7, 0x03, 0x02, 0x42,
+    0xdf, 0x54, 0x03, 0x64, 0x03, 0x70, 0x03, 0x02, 0x42, 0xcf, 0x7b, 0xcb,
+    0x12, 0x43, 0x2c, 0x8f, 0xfb, 0x7b, 0x30, 0x7d, 0x03, 0xe4, 0xff, 0x12,
+    0x3d, 0xd7, 0xc3, 0xef, 0x94, 0x02, 0xee, 0x94, 0x00, 0x50, 0x2a, 0x12,
+    0x42, 0xec, 0xef, 0x4e, 0x70, 0x23, 0x12, 0x43, 0x04, 0x60, 0x0a, 0x12,
+    0x43, 0x12, 0x70, 0x0c, 0x12, 0x43, 0x1f, 0x70, 0x07, 0x12, 0x46, 0x39,
+    0x7b, 0x03, 0x80, 0x07, 0x12, 0x46, 0x39, 0x12, 0x46, 0x43, 0xfb, 0x7a,
+    0x00, 0x7d, 0x54, 0x80, 0x3e, 0x12, 0x42, 0xec, 0xef, 0x4e, 0x70, 0x24,
+    0x12, 0x43, 0x04, 0x60, 0x0a, 0x12, 0x43, 0x12, 0x70, 0x0f, 0x12, 0x43,
+    0x1f, 0x70, 0x0a, 0x12, 0x46, 0x39, 0xe4, 0xfb, 0xfa, 0x7d, 0xee, 0x80,
+    0x1e, 0x12, 0x46, 0x39, 0x7b, 0x01, 0x7a, 0x00, 0x7d, 0xee, 0x80, 0x13,
+    0x12, 0x46, 0x39, 0x12, 0x46, 0x43, 0x54, 0x40, 0xfe, 0xc4, 0x13, 0x13,
+    0x54, 0x03, 0xfb, 0x7a, 0x00, 0x7d, 0xee, 0x12, 0x38, 0xbd, 0x7b, 0xff,
+    0x12, 0x43, 0x2c, 0xef, 0x4e, 0x70, 0x07, 0x74, 0x2a, 0x25, 0x19, 0xf8,
+    0xe4, 0xf6, 0x05, 0x19, 0xe5, 0x19, 0xc3, 0x94, 0x02, 0x50, 0x03, 0x02,
+    0x42, 0x15, 0x22, 0xe5, 0x19, 0x24, 0x17, 0xfd, 0x7b, 0x20, 0x7f, 0x04,
+    0x12, 0x3d, 0xd7, 0x22, 0xe5, 0x19, 0x24, 0x17, 0xfd, 0x7f, 0x04, 0x12,
+    0x3d, 0xd7, 0x22, 0x7b, 0x22, 0x7d, 0x18, 0x7f, 0x06, 0x12, 0x3d, 0xd7,
+    0xef, 0x64, 0x01, 0x4e, 0x22, 0x7d, 0x1c, 0xe4, 0xff, 0x12, 0x3e, 0x9a,
+    0xef, 0x54, 0x1b, 0x64, 0x0a, 0x22, 0x7b, 0xcc, 0x7d, 0x10, 0xff, 0x12,
+    0x3d, 0xd7, 0xef, 0x64, 0x01, 0x4e, 0x22, 0xe5, 0x19, 0x24, 0x17, 0xfd,
+    0x7f, 0x04, 0x12, 0x3d, 0xd7, 0x22, 0xd2, 0x08, 0x75, 0xfb, 0x03, 0xab,
+    0x7e, 0xaa, 0x7d, 0x7d, 0x19, 0x7f, 0x03, 0x12, 0x3e, 0xda, 0xe5, 0x7e,
+    0x54, 0x0f, 0x24, 0xf3, 0x60, 0x03, 0x02, 0x43, 0xe9, 0x12, 0x46, 0x5a,
+    0x12, 0x46, 0x61, 0xd8, 0xfb, 0xff, 0x20, 0xe2, 0x35, 0x13, 0x92, 0x0c,
+    0xef, 0xa2, 0xe1, 0x92, 0x0b, 0x30, 0x0c, 0x2a, 0xe4, 0xf5, 0x10, 0x7b,
+    0xfe, 0x12, 0x43, 0xff, 0xef, 0xc4, 0x33, 0x33, 0x54, 0xc0, 0xff, 0xc0,
+    0x07, 0x7b, 0x54, 0x12, 0x43, 0xff, 0xd0, 0xe0, 0x4f, 0xff, 0x74, 0x2a,
+    0x25, 0x10, 0xf8, 0xa6, 0x07, 0x05, 0x10, 0xe5, 0x10, 0xc3, 0x94, 0x02,
+    0x40, 0xd9, 0x12, 0x46, 0x5a, 0x12, 0x46, 0x61, 0xd8, 0xfb, 0x54, 0x05,
+    0x64, 0x04, 0x70, 0x27, 0x78, 0xc4, 0xe6, 0x78, 0xc6, 0xf6, 0xe5, 0x7d,
+    0xff, 0x33, 0x95, 0xe0, 0xef, 0x54, 0x0f, 0x78, 0xc4, 0xf6, 0x12, 0x44,
+    0x0a, 0x20, 0x0c, 0x0c, 0x12, 0x46, 0x5a, 0x12, 0x46, 0x61, 0xd8, 0xfb,
+    0x13, 0x92, 0x0d, 0x22, 0xc2, 0x0d, 0x22, 0x12, 0x46, 0x5a, 0x12, 0x46,
+    0x61, 0xd8, 0xfb, 0x54, 0x05, 0x64, 0x05, 0x70, 0x1e, 0x78, 0xc4, 0x7d,
+    0xb8, 0x12, 0x43, 0xf5, 0x78, 0xc1, 0x7d, 0x74, 0x12, 0x43, 0xf5, 0xe4,
+    0x78, 0xc1, 0xf6, 0x22, 0x7b, 0x01, 0x7a, 0x00, 0x7d, 0xee, 0x7f, 0x92,
+    0x12, 0x38, 0xbd, 0x22, 0xe6, 0xfb, 0x7a, 0x00, 0x7f, 0x92, 0x12, 0x38,
+    0xbd, 0x22, 0xe5, 0x10, 0x24, 0x17, 0xfd, 0x7f, 0x04, 0x12, 0x3d, 0xd7,
+    0x22, 0x78, 0xc1, 0xe6, 0xfb, 0x7a, 0x00, 0x7d, 0x74, 0x7f, 0x92, 0x12,
+    0x38, 0xbd, 0xe4, 0x78, 0xc1, 0xf6, 0xf5, 0x11, 0x74, 0x01, 0x7e, 0x00,
+    0xa8, 0x11, 0x08, 0x80, 0x05, 0xc3, 0x33, 0xce, 0x33, 0xce, 0xd8, 0xf9,
+    0xff, 0x78, 0xc4, 0xe6, 0xfd, 0xef, 0x5d, 0x60, 0x44, 0x85, 0x11, 0xfb,
+    0xe5, 0x11, 0x54, 0x02, 0x25, 0xe0, 0x25, 0xe0, 0xfe, 0xe4, 0x24, 0x5b,
+    0xfb, 0xee, 0x12, 0x45, 0xed, 0x12, 0x3e, 0xda, 0x7b, 0x40, 0x7d, 0x11,
+    0x7f, 0x07, 0x12, 0x3d, 0xd7, 0x74, 0xc7, 0x25, 0x11, 0xf8, 0xa6, 0x07,
+    0x7b, 0x11, 0x7d, 0x12, 0x7f, 0x07, 0x12, 0x3d, 0xd7, 0xef, 0x4e, 0x60,
+    0x09, 0x74, 0xe7, 0x25, 0x11, 0xf8, 0x76, 0x04, 0x80, 0x07, 0x74, 0xe7,
+    0x25, 0x11, 0xf8, 0x76, 0x0a, 0x05, 0x11, 0xe5, 0x11, 0xc3, 0x94, 0x04,
+    0x40, 0x9a, 0x78, 0xc6, 0xe6, 0x70, 0x15, 0x78, 0xc4, 0xe6, 0x60, 0x10,
+    0x75, 0xd9, 0x38, 0x75, 0xdb, 0x10, 0x7d, 0xfe, 0x12, 0x44, 0xb8, 0x7d,
+    0x76, 0x12, 0x44, 0xb8, 0x79, 0xc6, 0xe7, 0x78, 0xc4, 0x66, 0xff, 0x60,
+    0x03, 0x12, 0x40, 0x25, 0x78, 0xc4, 0xe6, 0x70, 0x09, 0xfb, 0xfa, 0x7d,
+    0xfe, 0x7f, 0x8e, 0x12, 0x38, 0xbd, 0x22, 0x7b, 0x01, 0x7a, 0x00, 0x7f,
+    0x8e, 0x12, 0x38, 0xbd, 0x22, 0xe4, 0xf5, 0xfb, 0x7d, 0x1c, 0xe4, 0xff,
+    0x12, 0x3e, 0x9a, 0xad, 0x07, 0xac, 0x06, 0xec, 0x54, 0xc0, 0xff, 0xed,
+    0x54, 0x3f, 0x4f, 0xf5, 0x20, 0x30, 0x06, 0x2c, 0x30, 0x01, 0x08, 0xa2,
+    0x04, 0x72, 0x03, 0x92, 0x07, 0x80, 0x21, 0x30, 0x04, 0x06, 0x7b, 0xcc,
+    0x7d, 0x11, 0x80, 0x0d, 0x30, 0x03, 0x06, 0x7b, 0xcc, 0x7d, 0x10, 0x80,
+    0x04, 0x7b, 0x66, 0x7d, 0x16, 0xe4, 0xff, 0x12, 0x3d, 0xd7, 0xee, 0x4f,
+    0x24, 0xff, 0x92, 0x07, 0xaf, 0xfb, 0x74, 0x26, 0x2f, 0xf8, 0xe6, 0xff,
+    0xa6, 0x20, 0x20, 0x07, 0x39, 0x8f, 0x20, 0x30, 0x07, 0x34, 0x30, 0x00,
+    0x31, 0x20, 0x04, 0x2e, 0x20, 0x03, 0x2b, 0xe4, 0xf5, 0xff, 0x75, 0xfc,
+    0xc2, 0xe5, 0xfc, 0x30, 0xe0, 0xfb, 0xaf, 0xfe, 0xef, 0x20, 0xe3, 0x1a,
+    0xae, 0xfd, 0x44, 0x08, 0xf5, 0xfe, 0x75, 0xfc, 0x80, 0xe5, 0xfc, 0x30,
+    0xe0, 0xfb, 0x8f, 0xfe, 0x8e, 0xfd, 0x75, 0xfc, 0x80, 0xe5, 0xfc, 0x30,
+    0xe0, 0xfb, 0x05, 0xfb, 0xaf, 0xfb, 0xef, 0xc3, 0x94, 0x04, 0x50, 0x03,
+    0x02, 0x44, 0xc5, 0xe4, 0xf5, 0xfb, 0x22, 0xe5, 0x7e, 0x54, 0x0f, 0x64,
+    0x01, 0x70, 0x23, 0xe5, 0x7e, 0x30, 0xe4, 0x1e, 0x90, 0x47, 0xd0, 0xe0,
+    0x44, 0x02, 0xf0, 0x54, 0xfb, 0xf0, 0x90, 0x47, 0xd4, 0xe0, 0x44, 0x04,
+    0xf0, 0x7b, 0x03, 0x7d, 0x5b, 0x7f, 0x5d, 0x12, 0x36, 0x29, 0x7b, 0x0e,
+    0x80, 0x1c, 0x90, 0x47, 0xd0, 0xe0, 0x54, 0xfd, 0xf0, 0x44, 0x04, 0xf0,
+    0x90, 0x47, 0xd4, 0xe0, 0x54, 0xfb, 0xf0, 0x7b, 0x02, 0x7d, 0x5b, 0x7f,
+    0x5d, 0x12, 0x36, 0x29, 0x7b, 0x06, 0x7d, 0x60, 0x7f, 0x63, 0x12, 0x36,
+    0x29, 0x22, 0xe5, 0x7e, 0x30, 0xe5, 0x35, 0x30, 0xe4, 0x0b, 0x7b, 0x02,
+    0x7d, 0x33, 0x7f, 0x35, 0x12, 0x36, 0x29, 0x80, 0x10, 0x7b, 0x01, 0x7d,
+    0x33, 0x7f, 0x35, 0x12, 0x36, 0x29, 0x90, 0x47, 0xd2, 0xe0, 0x44, 0x04,
+    0xf0, 0x90, 0x47, 0xd2, 0xe0, 0x54, 0xf7, 0xf0, 0x90, 0x47, 0xd1, 0xe0,
+    0x44, 0x10, 0xf0, 0x7b, 0x05, 0x7d, 0x84, 0x7f, 0x86, 0x12, 0x36, 0x29,
+    0x22, 0xfb, 0xe5, 0x1c, 0x34, 0xf0, 0xfa, 0x7d, 0x10, 0x7f, 0x07, 0x22,
+    0x54, 0x01, 0xc4, 0x33, 0x54, 0xe0, 0xf5, 0xdb, 0x44, 0x08, 0xf5, 0xdb,
+    0x22, 0xf5, 0xdb, 0x75, 0xdb, 0x08, 0xf5, 0xdb, 0x75, 0xdb, 0x08, 0x22,
+    0xab, 0x07, 0xaa, 0x06, 0x7d, 0x10, 0x7f, 0x07, 0x12, 0x3e, 0xda, 0x7b,
+    0xff, 0x7d, 0x10, 0x7f, 0x07, 0x12, 0x3d, 0xd7, 0xef, 0x4e, 0x60, 0xf3,
+    0x22, 0x12, 0x44, 0xc2, 0x30, 0x0c, 0x03, 0x12, 0x42, 0x12, 0x78, 0xc4,
+    0xe6, 0xff, 0x60, 0x03, 0x12, 0x40, 0x25, 0x22, 0xe5, 0x19, 0x24, 0x17,
+    0x54, 0x1f, 0x44, 0x80, 0xff, 0x22, 0x74, 0x2a, 0x25, 0x19, 0xf8, 0xe6,
+    0x22, 0x12, 0x46, 0x72, 0x12, 0x46, 0x68, 0x90, 0x47, 0xfa, 0xe0, 0x54,
+    0xf8, 0x44, 0x02, 0xf0, 0x22, 0xe5, 0x7e, 0xae, 0x7d, 0x78, 0x04, 0x22,
+    0xce, 0xa2, 0xe7, 0x13, 0xce, 0x13, 0x22, 0xe4, 0x78, 0xc4, 0xf6, 0xc2,
+    0x0d, 0x78, 0xc1, 0xf6, 0x22, 0xc2, 0x0c, 0xc2, 0x0b, 0x22, 0x22,
+];
